chore: run wizard to rename project and author
