+++ README.md
@@ -1,7 +1,7 @@
 # The Wizard Template for Python Projects
 
 <p align="center">
-<a href="https://github.com/fschuch/wizard-template"><img src="https://raw.githubusercontent.com/fschuch/wizard-template/refs/heads/main/docs/logo.png" alt="Wizard template logo" width="320"></a>
+<a href="https://github.com/fschuch/xarray-binfile"><img src="https://raw.githubusercontent.com/fschuch/xarray-binfile/refs/heads/main/docs/logo.png" alt="Wizard template logo" width="320"></a>
 </p>
 <p align="center">
     <em>Let the wizard do the heavy lifting so you can focus on your craft</em>
@@ -10,17 +10,17 @@
 ______________________________________________________________________
 
 - CI/CD:
-  [![CI](https://github.com/fschuch/wizard-template/actions/workflows/ci.yaml/badge.svg?branch=main)](https://github.com/fschuch/wizard-template/actions/workflows/ci.yaml)
-  [![Docs](https://github.com/fschuch/wizard-template/actions/workflows/docs.yaml/badge.svg?branch=main)](https://docs.fschuch.com/wizard-template)
-  [![CodeQL](https://github.com/fschuch/wizard-template/actions/workflows/github-code-scanning/codeql/badge.svg)](https://github.com/fschuch/wizard-template/actions/workflows/github-code-scanning/codeql)
-  [![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=fschuch_wizard-template&metric=alert_status)](https://sonarcloud.io/summary/new_code?id=fschuch_wizard-template)
+  [![CI](https://github.com/fschuch/xarray-binfile/actions/workflows/ci.yaml/badge.svg?branch=main)](https://github.com/fschuch/xarray-binfile/actions/workflows/ci.yaml)
+  [![Docs](https://github.com/fschuch/xarray-binfile/actions/workflows/docs.yaml/badge.svg?branch=main)](https://docs.fschuch.com/xarray-binfile)
+  [![CodeQL](https://github.com/fschuch/xarray-binfile/actions/workflows/github-code-scanning/codeql/badge.svg)](https://github.com/fschuch/xarray-binfile/actions/workflows/github-code-scanning/codeql)
+  [![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=fschuch_xarray-binfile&metric=alert_status)](https://sonarcloud.io/summary/new_code?id=fschuch_xarray-binfile)
 
 - Meta:
   [![Wizard Template](https://img.shields.io/badge/Wizard-Template-%23447CAA)](https://github.com/fschuch/wizard-template)
   [![Checked with mypy](https://www.mypy-lang.org/static/mypy_badge.svg)](https://mypy-lang.org/)
   [![Hatch project](https://img.shields.io/badge/%F0%9F%A5%9A-Hatch-4051b5.svg)](https://github.com/pypa/hatch)
   [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
-  ![GitHub License](https://img.shields.io/github/license/fschuch/wizard-template?color=blue)
+  ![GitHub License](https://img.shields.io/github/license/fschuch/xarray-binfile?color=blue)
   [![EffVer Versioning](https://img.shields.io/badge/version_scheme-EffVer-0097a7)](https://jacobtomlinson.dev/effver)
 
 ______________________________________________________________________
@@ -29,9 +29,9 @@ ______________________________________________________________________
 
 This is a general-purpose template that aims to provide a magical start to any Python project. It includes the initial configuration of quality assurance tools, documentation, and automated actions to deploy a Python package.
 
-Check out the documentation for more details on how to use the template and its features: <https://docs.fschuch.com/wizard-template>.
+Check out the documentation for more details on how to use the template and its features: <https://docs.fschuch.com/xarray-binfile>.
 
 ## Copyright and License
 
 © 2023 [Felipe N. Schuch](https://github.com/fschuch).
-All content is under [MIT License](https://github.com/fschuch/wizard-template/blob/main/LICENSE).
+All content is under [MIT License](https://github.com/fschuch/xarray-binfile/blob/main/LICENSE).
