+++ docs/references/what-is-new.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Release notes are automatically recovered from GitHub by [sphinx-github-changelog](https://github.com/ewjoachim/sphinx-github-changelog).
 
 ```{changelog}
-:changelog-url: https://docs.fschuch.com/wizard-template/references/what-is-new.html
-:github: https://github.com/fschuch/wizard-template/releases/
-:pypi: https://pypi.org/project/wizard-template/
+:changelog-url: https://docs.fschuch.com/xarray-binfile/references/what-is-new.html
+:github: https://github.com/fschuch/xarray-binfile/releases/
+:pypi: https://pypi.org/project/xarray-binfile/
 ```
