+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = ["hatchling>=1.21.0", "hatch-vcs>=0.3.0"]
 build-backend = "hatchling.build"
 
 [project]
-name = "wizard-template"
+name = "xarray-binfile"
 description = "A template for a python project containing a package, tests, docs, and CI/CD."
 readme = "README.md"
 license = "MIT"
@@ -46,10 +46,10 @@ docs = [
 ]
 
 [project.urls]
-Source = "https://github.com/fschuch/wizard-template"
-Tracker = "https://github.com/fschuch/wizard-template/issues"
-Changelog = "https://fschuch.github.io/wizard-template/references/what-is-new.html"
-Documentation = "https://fschuch.github.io/wizard-template/"
+Source = "https://github.com/fschuch/xarray-binfile"
+Tracker = "https://github.com/fschuch/xarray-binfile/issues"
+Changelog = "https://fschuch.github.io/xarray-binfile/references/what-is-new.html"
+Documentation = "https://fschuch.github.io/xarray-binfile/"
 
 [project.scripts]
 
@@ -64,8 +64,8 @@ minversion = "8.0"
 [tool.coverage.run]
 branch = true
 relative_files = true
-source = ["src/wizard_template", "tests"]
-omit = ["src/wizard_template/_version.py"]
+source = ["src/xarray_binfile", "tests"]
+omit = ["src/xarray_binfile/_version.py"]
 
 [tool.coverage.report]
 fail_under = 90
@@ -96,7 +96,7 @@ check-filenames = true
 source = "vcs"
 
 [tool.hatch.build.hooks.vcs]
-version-file = "src/wizard_template/_version.py"
+version-file = "src/xarray_binfile/_version.py"
 
 [tool.hatch.envs.default]
 description = "Base development environment"
@@ -104,7 +104,6 @@ features = ["tests"]
 installer = "uv"
 
 [tool.hatch.envs.default.scripts]
-_wizard = ["python ./tools/rename_project_content.py"]
 pre-commit-install = "pre-commit install {args}"
 pre-commit-uninstall = "pre-commit uninstall {args}"
 check = "pre-commit run {args} --all-files"
@@ -137,5 +136,5 @@ config = "jupyter-book config sphinx docs {args}"
 build = ["config", "jupyter-book build docs --path-output build {args}"]
 serve = [
     "config",
-    "sphinx-autobuild docs build/_build/html --open-browser --watch src/wizard_template {args}",
+    "sphinx-autobuild docs build/_build/html --open-browser --watch src/xarray_binfile {args}",
 ]
