--- src/wizard_template/__init__.py
@@ -1,4 +1,0 @@
-"""Blueprint for a quick start of a new Python project."""
-
-from wizard_template._version import __version__
-from wizard_template.core import demo_function
