+++ src/xarray_binfile/__init__.py
@@ -1,0 +1,4 @@
+"""Blueprint for a quick start of a new Python project."""
+
+from xarray_binfile._version import __version__
+from xarray_binfile.core import demo_function
