+++ src/xarray_binfile/core.py
